Removes internal tests
