+++ internal/db/db_private_keys_test.go
@@ -1,10 +1,12 @@
 package db_test
 
 import (
+	"database/sql"
 	"path/filepath"
 	"testing"
 
 	"github.com/canonical/notary/internal/db"
+	"github.com/canonical/notary/internal/encryption"
 )
 
 func TestPrivateKeysEndToEnd(t *testing.T) {
@@ -84,5 +86,54 @@ func TestPrivateKeyFails(t *testing.T) {
 	_, err = database.GetDecryptedPrivateKey(db.ByPrivateKeyID(10))
 	if err == nil {
 		t.Fatalf("Should have failed to get private key")
+	}
+}
+
+func TestPrivateKeyEncryption(t *testing.T) {
+	tempDir := t.TempDir()
+	databasePath := filepath.Join(tempDir, "db.sqlite3")
+	database, err := db.NewDatabase(databasePath)
+	if err != nil {
+		t.Fatalf("Couldn't complete NewDatabase: %s", err)
+	}
+	defer database.Close()
+
+	pkID, err := database.CreatePrivateKey(RootCAPrivateKey)
+	if err != nil {
+		t.Fatalf("Couldn't create private key: %s", err)
+	}
+
+	pk := db.PrivateKey{PrivateKeyID: pkID}
+	sqlConnection, err := sql.Open("sqlite3", databasePath)
+	if err != nil {
+		t.Fatalf("Couldn't open database: %s", err)
+	}
+	defer sqlConnection.Close()
+	row := sqlConnection.QueryRow("SELECT * FROM private_keys WHERE private_key_id = ?", pk.PrivateKeyID)
+	err = row.Scan(&pk.PrivateKeyID, &pk.PrivateKeyPEM)
+	if err != nil {
+		t.Fatalf("Couldn't query raw secret: %s", err)
+	}
+
+	if pk.PrivateKeyPEM == RootCAPrivateKey {
+		t.Fatal("Private key is stored in plaintext!")
+	}
+
+	decryptedPK, err := database.GetDecryptedPrivateKey(db.ByPrivateKeyID(pkID))
+	if err != nil {
+		t.Fatalf("Couldn't get private key: %s", err)
+	}
+	if decryptedPK.PrivateKeyPEM != RootCAPrivateKey {
+		t.Fatalf("Decrypted secret doesn't match original. Got %q, want %q",
+			decryptedPK.PrivateKeyPEM, RootCAPrivateKey)
+	}
+
+	decryptedManually, err := encryption.Decrypt(pk.PrivateKeyPEM, database.EncryptionKey)
+	if err != nil {
+		t.Fatalf("Couldn't manually decrypt secret: %s", err)
+	}
+	if decryptedManually != RootCAPrivateKey {
+		t.Fatalf("Manually decrypted secret doesn't match original. Got %q, want %q",
+			decryptedManually, RootCAPrivateKey)
 	}
 }
